add email validation
